+++ ccip-sdk/viem.ts
@@ -1,0 +1,5 @@
+/**
+ * Viem adapters entry point.
+ * Import from '@chainlink/ccip-sdk/viem'
+ */
+export * from './src/evm/viem/index.ts'